uninstall node deps, styling
